Added utils.py [AUTH-403]
