+++ opentrons-ai-server/api/domain/utils.py
@@ -1,0 +1,73 @@
+def refine_characters(prompt: str) -> str:
+    """
+    Converts specific Greek characters in a string to their English phonetic equivalents and replaces
+    certain special characters. The function is designed to handle text with Greek characters and
+    special characters like backticks, converting them into more standardized or readable forms while
+    preserving the structure and formatting of the original text.
+
+    Parameters:
+    - text (str): The input string containing Greek characters and possibly special characters.
+
+    Returns:
+    - str: The modified string with Greek characters replaced by their English phonetic equivalents
+        and certain special characters like backticks replaced with single quotes.
+
+    Example:
+    >>> refine_characters("Transfer `10μ`")
+    'Transfer '10m''
+    """
+
+    greek_to_english = {
+        "α": "a",
+        "β": "b",
+        "γ": "g",
+        "δ": "d",
+        "ε": "e",
+        "ζ": "z",
+        "η": "e",
+        "θ": "th",
+        "ι": "i",
+        "κ": "k",
+        "λ": "l",
+        "μ": "m",
+        "ν": "n",
+        "ξ": "x",
+        "ο": "o",
+        "π": "p",
+        "ρ": "r",
+        "σ": "s",
+        "ς": "s",
+        "τ": "t",
+        "υ": "u",
+        "φ": "ph",
+        "χ": "ch",
+        "ψ": "ps",
+        "ω": "o",
+        "Α": "A",
+        "Β": "B",
+        "Γ": "G",
+        "Δ": "D",
+        "Ε": "E",
+        "Ζ": "Z",
+        "Η": "E",
+        "Θ": "Th",
+        "Ι": "I",
+        "Κ": "K",
+        "Λ": "L",
+        "Μ": "M",
+        "Ν": "N",
+        "Ξ": "X",
+        "Ο": "O",
+        "Π": "P",
+        "Ρ": "R",
+        "Σ": "S",
+        "Τ": "T",
+        "Υ": "U",
+        "Φ": "Ph",
+        "Χ": "Ch",
+        "Ψ": "Ps",
+        "Ω": "O",
+    }
+    translation_table = str.maketrans(greek_to_english)
+    translated_text = prompt.translate(translation_table)
+    return translated_text
